Refactoring: extract process handling from async job.
