+++ bootstrap/helpers.php
@@ -1,7 +1,6 @@
 <?php
 
 use App\Services\CoolifyProcess;
-use Illuminate\Process\ProcessResult;
 use Spatie\Activitylog\Contracts\Activity;
 
 if (! function_exists('coolifyProcess')) {
@@ -10,7 +9,7 @@ if (! function_exists('coolifyProcess')) {
      * Run a Coolify Process, which SSH's into a machine to run the command(s).
      *
      */
-    function coolifyProcess($command, $destination): Activity|ProcessResult
+    function coolifyProcess($command, $destination): Activity
     {
         $process = resolve(CoolifyProcess::class, [
             'destination' => $destination,
